requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
argon2-cffi=20.1.0=py38h1e0a361_0
astroid=2.4.2=py38_0
attrs=19.3.0=py_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blas=1.0=mkl
bleach=3.1.5=pyh9f0ad1d_0
brotlipy=0.7.0=py38h1e0a361_1000
ca-certificates=2020.6.24=0
certifi=2020.6.20=py38_0
cffi=1.14.0=py38he30daa8_1
chardet=3.0.4=py38h32f6830_1006
cryptography=3.0=py38h766eaa4_0
cycler=0.10.0=py38_0
dbus=1.13.16=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py38h32f6830_1001
expat=2.2.9=he6710b0_2
fontconfig=2.13.0=h9420a91_0
freetype=2.10.2=h5ab3b9f_0
glib=2.65.0=h3eb4bd4_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=1.7.0=py38h32f6830_0
importlib_metadata=1.7.0=0
intel-openmp=2020.1=217
ipykernel=5.3.4=py38h23f93f0_0
ipython=7.17.0=py38h1cdfbd6_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.5.1=py_0
isort=4.3.21=py38_0
jedi=0.17.2=py38h32f6830_0
jinja2=2.11.2=pyh9f0ad1d_0
jpeg=9b=h024ee3a_2
json5=0.9.4=pyh9f0ad1d_0
jsonschema=3.2.0=py38h32f6830_1
jupyter=1.0.0=py38_7
jupyter_client=6.1.6=py_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.3=py38h32f6830_1
jupyterlab=2.2.2=py_0
jupyterlab_server=1.2.0=py_0
kiwisolver=1.2.0=py38hfd86e86_0
lazy-object-proxy=1.4.3=py38h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h516909a_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he6710b0_1
markupsafe=1.1.1=py38h1e0a361_1
matplotlib=3.2.2=0
matplotlib-base=3.2.2=py38hef1b27d_0
mccabe=0.6.1=py38_1
mistune=0.8.4=py38h1e0a361_1001
mkl=2020.1=217
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.1.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
nbconvert=5.6.1=py38h32f6830_1
nbformat=5.0.7=py_0
ncurses=6.2=he6710b0_1
notebook=6.1.0=py38h32f6830_0
numpy=1.19.1=py38hbc911f0_0
numpy-base=1.19.1=py38hfa32c7d_0
olefile=0.46=py_0
openssl=1.1.1g=h7b6447c_0
packaging=20.4=pyh9f0ad1d_0
pandoc=2.10.1=h516909a_0
pandocfilters=1.4.2=py_1
parso=0.7.1=pyh9f0ad1d_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py38h32f6830_1
pickleshare=0.7.5=py38h32f6830_1001
pillow=7.2.0=py38hb39fc2d_0
pip=20.1.1=py38_1
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt-toolkit=3.0.5=py_1
prompt_toolkit=3.0.5=0
ptyprocess=0.6.0=py_1001
pycparser=2.20=pyh9f0ad1d_2
pygments=2.6.1=py_0
pylint=2.5.3=py38_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.16.0=py38h1e0a361_0
pysocks=1.7.1=py38h32f6830_1
python=3.8.3=hcff3b4d_2
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pyzmq=19.0.2=py38ha71036d_0
qt=5.9.7=h5867ecd_1
qtconsole=4.7.5=py_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
requests=2.24.0=pyh9f0ad1d_0
send2trash=1.5.0=py_0
setuptools=49.2.0=py38_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=pyh9f0ad1d_0
sqlite=3.32.3=h62c20be_0
terminado=0.8.3=py38h32f6830_1
testpath=0.4.4=py_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
tornado=6.0.4=py38h1e0a361_1
traitlets=4.3.3=py38h32f6830_1
urllib3=1.25.10=py_0
wcwidth=0.2.5=pyh9f0ad1d_1
webencodings=0.5.1=py_1
wheel=0.34.2=py38_0
widgetsnbextension=3.5.1=py38_0
wrapt=1.11.2=py38h7b6447c_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.2=he1b5a44_3
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
